Fix compilation errors causing Vercel build failure
- public_home.dart: remove stray extra ')' causing parse error (line 1021)
- doctor_consultation_screen.dart: fix int→double type mismatch for VitalSigns fields
- doctor_referrals_list.dart: remove const keyword and fix .reason→.reasonForReferral
- create_referral_screen.dart: make patient field nullable (User?)
- patient_profile.dart: fix logout()→setUserLogout() method name

+++ lib/screens/doctor_consultation_screen.dart
@@ -1370,9 +1370,9 @@ class _DoctorConsultationScreenState
         ),
         examination: PhysicalExamination(
           vitalSigns: VitalSigns(
-            bloodPressureSystolic: int.tryParse(_bpSystolicController.text) ?? 0,
-            bloodPressureDiastolic: int.tryParse(_bpDiastolicController.text) ?? 0,
-            heartRate: int.tryParse(_heartRateController.text) ?? 0,
+            bloodPressureSystolic: double.tryParse(_bpSystolicController.text) ?? 0,
+            bloodPressureDiastolic: double.tryParse(_bpDiastolicController.text) ?? 0,
+            heartRate: double.tryParse(_heartRateController.text) ?? 0,
             temperature: double.tryParse(_temperatureController.text) ?? 0,
             oxygenSaturation: double.tryParse(_oxygenSaturationController.text) ?? 0,
             weight: double.tryParse(_weightController.text) ?? 0,
